Average of three numbers

+++ AverageOfThreeNumbers.java
@@ -3,11 +3,11 @@ public class AverageOfThreeNumbers{
 public static void main(String []args){
 Scanner input=new Scanner(System.in);
 System.out.println("Enter the first number :");
-double a=input.nextInt();
+double a=input.nextDouble();
 System.out.println("Enter the second number :");
-double b=input.nextInt();
+double b=input.nextDouble();
 System.out.println("Enter the third number :");
-double c=input.nextInt();
+double c=input.nextDouble();
 System.out.print("Average of three numbers is :"+(a+b+c)/3);
 }
 }
